Address a rare issue with query multiple results
Occasionally, it was observed that a stored procedure with multiple
statements would fail to return results for some of them.

Although virtually impossible to confirm, we hypothesize that this is
due to the `MySQL2::Result` objects being accessed before they have time
to be fully realized due to the asynchronous nature of MySQL.

This PR waits until all results have been fully gathered before mapping
the objects to arrays, allowing time for the stored proc to execute
fully.

Previously, we were observing a failure rate of approx 1%. With this fix
we have run 100,000 iterations of the same test, resulting in 0
failures.