Adds support for new SQL clauses.
Added increment and decrement methods that can be used with or without
an existing set clause. These methods are intended to be used to
increment or decrement an integer column entirely on the server.

Also added support for offset and having clauses and tidied up the
generated SQL a little.

Signed-off-by: Rob Kilby <rob.kilby@sage.com>