Added blocking atomic locks around migrations.
This is intended to ease the situation where multiple instances of an
application are running, and are each separately trying to apply
migrations during startup.

Currently, this would result in an Exception being raised when a lock
could not be obtained - this implementation blocks (within a defined
retry limit) until the lock is either released or becomes available in
some other way.

Signed-off-by: Rob Kilby <rob.kilby@sage.com>