Fix script manager lock key
mysql_framework depends on `redlock`, without any version restriction.

However, starting with 2.0.0, `redlock` changed its implementation
to use the gem `redis-client`, which does not allow passing in a
`Class` type as a locking key.

This commit addresses this by using the class name instead. The previous
`redis` implementation was calling `#to_s` on the argument, so the end
result is the same.